backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
beautifulsoup4==4.12.3
openpyxl==3.1.2
python-multipart==0.0.9
python-dateutil==2.9.0
